requirements.txt
geographiclib==1.50
geopy==2.0.0
python-telegram-bot==12.2.0
requests==2.23.0
urllib3==1.25.8
python-dotenv==0.14.0
emoji
numpy>=1.17.3
pandas
plotly
sklearn
google-api-python-client
google-auth-httplib2
google-auth-oauthlib